+++ pallets/swap/src/mock.rs
@@ -83,7 +83,7 @@ parameter_types! {
     pub const SwapProtocolId: PalletId = PalletId(*b"ten/swap");
     pub const MaxFeeRate: u16 = 10000; // 15.26%
     pub const MaxPositions: u32 = 100;
-    pub const MinimumLiquidity: u64 = 1_000;
+    pub const MinimumLiquidity: u128 = 1_000;
     pub const MinimumReserves: NonZeroU64 = NonZeroU64::new(1).unwrap();
 }
