+++ runtime/src/lib.rs
@@ -1296,7 +1296,7 @@ parameter_types! {
     pub const SwapProtocolId: PalletId = PalletId(*b"ten/swap");
     pub const SwapMaxFeeRate: u16 = 10000; // 15.26%
     pub const SwapMaxPositions: u32 = 100;
-    pub const SwapMinimumLiquidity: u64 = 1_000;
+    pub const SwapMinimumLiquidity: u128 = 1_000;
     pub const SwapMinimumReserve: NonZeroU64 = unsafe { NonZeroU64::new_unchecked(1_000_000) };
 }
